requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
httpx==0.27.0
fastapi>=0.100.0
uvicorn>=0.22.0
